+++ packages/widgetbook/test/generator/generic_bound/generic_bound.stories.g.dart
@@ -25,11 +25,13 @@ class BoundWidgetStory<D, T extends BaseItem<D>>
     extends Story<BoundWidget<D, T>, BoundWidgetArgs<D, T>> {
   BoundWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     required super.args,
     StoryWidgetBuilder<BoundWidget<D, T>, BoundWidgetArgs<D, T>>? builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          builder:
              builder ??