+++ packages/widgetbook/test/generator/generic_bound_nullable/generic_bound_nullable.stories.g.dart
@@ -29,6 +29,7 @@ class NullableBoundWidgetStory<D, T extends BaseItem<D?>>
     extends Story<NullableBoundWidget<D, T>, NullableBoundWidgetArgs<D, T>> {
   NullableBoundWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     required super.args,
@@ -38,6 +39,7 @@ class NullableBoundWidgetStory<D, T extends BaseItem<D?>>
     >?
     builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          builder:
              builder ??