+++ packages/widgetbook/test/generator/static_methods/static_methods.stories.g.dart
@@ -28,11 +28,13 @@ class StaticMethodWidgetStory
     extends Story<StaticMethodWidget, StaticMethodWidgetArgs> {
   StaticMethodWidgetStory({
     super.name,
+    super.designLink,
     super.setup,
     super.modes,
     StaticMethodWidgetArgs? args,
     StoryWidgetBuilder<StaticMethodWidget, StaticMethodWidgetArgs>? builder,
     super.scenarios,
+    super.excludeFromTests,
   }) : super(
          args: args ?? StaticMethodWidgetArgs(),
          builder: